feat: BlockUsers 테이블 추가 및 Change ID columns Type Int to Long

+++ src/main/kotlin/com/coffeetime/ctauth/domain/model/UserRegisterResponse.kt
@@ -1,9 +1,9 @@
 package com.coffeetime.ctauth.domain.model
 
 data class UserRegisterResponse (
-    var userID: Int
+    var userID: Long
 ) {
-    fun userID(userID: Int): UserRegisterResponse {
+    fun userID(userID: Long): UserRegisterResponse {
         this.userID = userID
         return this
     }
